Seaprate time tag and time tags utils.
